requirements.txt
streamlit
python-dotenv
groq
streamlit-option-menu
pandas
numpy
plotly
streamlit-lottie
requests
sounddevice
